requirements.txt
# FastAPI dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pydantic[email]==2.12.5
python-multipart>=0.0.6

# Environment variables
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.45
pyodbc>=5.3.0
aioodbc>=0.5.0
alembic>=1.13.0

# Authentication & Security
passlib==1.7.4
bcrypt==4.0.1
PyJWT==2.10.1
python-dateutil>=2.8.2
# Observability - Metrics, Logging, and Tracing
prometheus-client>=0.19.0
prometheus-fastapi-instrumentator>=6.1.0
python-logging-loki>=0.3.1
opentelemetry-api>=1.28.2
opentelemetry-sdk>=1.28.2
opentelemetry-instrumentation-fastapi>=0.49b2
opentelemetry-instrumentation-sqlalchemy>=0.49b2
opentelemetry-exporter-otlp-proto-grpc>=1.28.2

# Testing
pytest>=8.0
pytest-asyncio>=0.24